requirements.txt
flask==3.0.2
flask-cors==4.0.0
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
loguru==0.7.2
requests==2.31.0
# Gemma/Google Cloud dependencies (replacing Groq)
google-cloud-aiplatform>=1.35.0
vertexai>=1.35.0
openai>=1.0.0
google-auth>=2.17.0
# LangChain and graph dependencies
langchain==0.1.4
langchain-community==0.0.19
langchain-core==0.1.22
langgraph==0.1.1
langgraph-checkpoint==1.0.2
# Other dependencies
python-dateutil==2.8.2
pytz==2024.1
retry==0.9.2
streamlit==1.29.0 